requirements.txt
asgiref==3.4.1
Django==3.2.9
django-debug-toolbar==3.2.2
django-redis==5.1.0
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
